feat: Upgrade to React 19 and latest Storybook version
This update modernizes the project by upgrading key dependencies and migrating components to improve maintainability. Upgraded  and  to version 19. Updated  to support React versions 16 through 19. Migrated  from a class component to a functional component with hooks. Upgraded Storybook from version 5 to 9, which is compatible with React 19 and offers improved documentation capabilities. Modified  scripts to align with the new Storybook version.